add fluent vaidators alergia controller

+++ MedicalRecord_API/Controllers/AlergiaController.cs
@@ -1,14 +1,12 @@
 ﻿using AutoMapper;
-using MedicalRecord_API.Repository.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using MedicalRecord_API.Utils.Response;
 using MedicalRecord_API.Models.Dtos.Alergia;
 using MedicalRecord_API.Models;
 using System.Net;
-using Microsoft.AspNetCore.Authorization;
 using MedicalRecord_API.Services.Interfaces;
-using MedicalRecord_API.Models.Dtos.Cie;
-
+using FluentValidation.Results;
+using MedicalRecord_API.Validators.Alergia;
 namespace MedicalRecord_API.Controllers
 {
     [ApiController]
@@ -40,7 +38,7 @@ namespace MedicalRecord_API.Controllers
             catch
             {
                 _response.Status = HttpStatusCode.InternalServerError;
-                _response.ErrorMessages = ["Ocurrió un error al procesar la solicitud"];
+                _response.ErrorMessages = ["Error al procesar la solicitud en el servidor."];
                 return StatusCode(StatusCodes.Status500InternalServerError, _response);
             }
         }
@@ -57,7 +55,7 @@ namespace MedicalRecord_API.Controllers
                 if (dto == null)
                 {
                     _response.Status = HttpStatusCode.NotFound;
-                    _response.ErrorMessages = [" modelo: no esxiste en la base de datos"];
+                    _response.ErrorMessages = ["Sin registros para este identificador"];
                     return NotFound(_response);
                 }
                 _response.Status = HttpStatusCode.OK;
@@ -68,7 +66,7 @@ namespace MedicalRecord_API.Controllers
             catch
             {
                 _response.Status = HttpStatusCode.InternalServerError;
-                _response.ErrorMessages = ["Ocurrió un error al procesar la solicitud"];
+                _response.ErrorMessages = ["Error al procesar la solicitud en el servidor."];
                 return StatusCode(StatusCodes.Status500InternalServerError, _response);
             }
         }
@@ -78,19 +76,34 @@ namespace MedicalRecord_API.Controllers
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<Response>> Create([FromBody] AlergiaCreateDto dto)
         {
+            AlergiaCreateDtoVaidator validator = new();
+
+            ValidationResult result =validator.Validate(dto);
+            if (!result.IsValid)
+            {
+                _response.Status = HttpStatusCode.BadRequest;
+                _response.ErrorMessages = result.Errors.Select(e => e.ErrorMessage).ToList();
+                return BadRequest(_response);
+            }
+            if(await _service.GetAsync(a=>a.Nombre==dto.Nombre,false)!=null)
+            {
+                _response.Status = HttpStatusCode.BadRequest;
+                _response.ErrorMessages = ["Ya existe un registro con este nombre"];
+                return BadRequest(_response);
+            }
             try
             {
                  Alergium modelo=_mapper.Map<Alergium>(dto);
                  modelo=await _service.Create(modelo);
                 _response.Status = HttpStatusCode.Created;
                 _response.IsSuccess = true;
-                _response.Result = _mapper.Map<AlergiaCreateDto>(modelo);
+                _response.Result = _mapper.Map<AlergiaDto>(modelo);
                 return CreatedAtRoute("GetAlergia", new { id = modelo.Id }, _response);
             }
             catch
             {
                 _response.Status = HttpStatusCode.InternalServerError;
-                _response.ErrorMessages = ["Ocurrió un error al procesar la solicitud."];
+                _response.ErrorMessages = ["Error al procesar la solicitud en el servidor."];
                 return StatusCode(StatusCodes.Status500InternalServerError, _response);
             }
         }
@@ -101,10 +114,18 @@ namespace MedicalRecord_API.Controllers
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<Response>> Update(int id, [FromBody] AlergiaUpdateDto dto)
         {
+            AlergiaUpdateDtoValidator validator = new();
+            ValidationResult result = validator.Validate(dto);
+            if (!result.IsValid)
+            {
+                _response.Status = HttpStatusCode.BadRequest;
+                _response.ErrorMessages = result.Errors.Select(e => e.ErrorMessage).ToList();
+                return BadRequest(_response);
+            }
             if (id != dto.Id)
             {
                 _response.Status = HttpStatusCode.BadRequest;
-                _response.ErrorMessages = ["El identificador no coinicde con el id del modelo"];
+                _response.ErrorMessages = ["El identificador es diferente al id del modelo"];
                 return BadRequest(_response);
             }
 
@@ -113,7 +134,7 @@ namespace MedicalRecord_API.Controllers
                 if (await _service.GetAsync(a=> a.Id==id,false)==null)
                 {
                     _response.Status = HttpStatusCode.NotFound;
-                    _response.ErrorMessages = ["Identificador no encontrado en los registros"];
+                    _response.ErrorMessages = ["Sin registros para este identificador"];
                     return NotFound(_response);
                 }
 
@@ -125,7 +146,7 @@ namespace MedicalRecord_API.Controllers
             catch
             {
                 _response.Status = HttpStatusCode.InternalServerError;
-                _response.ErrorMessages = ["Ocurrió un error al procesar la solicitud."];
+                _response.ErrorMessages = ["Error al procesar la solicitud en el servidor."];
                 return StatusCode(StatusCodes.Status500InternalServerError, _response);
             }
         }
@@ -137,13 +158,20 @@ namespace MedicalRecord_API.Controllers
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<Response>> Delete(int id)
         {
+            bool isValid=id>0;
+            if (!isValid)
+            {
+                _response.Status = HttpStatusCode.BadRequest;
+                _response.ErrorMessages = ["El identificador no es valido"];
+                return BadRequest(_response);
+            }
             try
             {
                 Alergium alergia =await _service.GetAsync(a => a.Id == id);
                 if (alergia==null)
                 {
                     _response.Status = HttpStatusCode.NotFound;
-                    _response.ErrorMessages = ["este identificador ya no existe en los registros"];
+                    _response.ErrorMessages = ["Sin registros para este identificador"];
                     return NotFound(_response);
                 }
                 await _service.Delete(alergia);
@@ -154,7 +182,7 @@ namespace MedicalRecord_API.Controllers
             catch
             {
                 _response.Status = HttpStatusCode.InternalServerError;
-                _response.ErrorMessages = ["Ocurrió un error al procesar la solicitud."];
+                _response.ErrorMessages = ["Error al procesar la solicitud en el servidor."];
                 return StatusCode(StatusCodes.Status500InternalServerError, _response);
             }
         }
